refactor: rename contentscanner to content-scanner and update imports
